requirements.txt
discord.py[voice]==2.0.1
